more test in the right locations
