--- app/views/old_layouts/_header_navigation.html.erb
@@ -1,18 +1,0 @@
-<% if signed_in? %>
-  <nav aria-label="<%= t('.nav_aria_label') %>" class="govuk-header__navigation">
-    <button type="button" class="govuk-header__menu-button govuk-js-header-toggle" aria-controls="navigation" aria-label="<%= t('.button_aria_label') %>" hidden>
-      <%= t('.menu_button') %>
-    </button>
-    <ul id="navigation" class="govuk-header__navigation-list app-header-menu-hidden-on-load">
-      <li class="govuk-header__navigation-item">
-        <%= link_to t('.your_applications'), claims_path, class: 'govuk-header__link' %>
-      </li>
-      <li class="govuk-header__navigation-item app-header__auth-user">
-        <%= current_provider.display_name %>
-      </li>
-      <li class="govuk-header__navigation-item">
-        <%= button_to t('.sign_out'), providers_logout_path, method: :delete, class: 'govuk-header__link app-header__auth-link' %>
-      </li>
-    </ul>
-  </nav>
-<% end %>
