--- app/views/old_layouts/_phase_banner.html.erb
@@ -1,8 +1,0 @@
-<div class="govuk-phase-banner">
-  <p class="govuk-phase-banner__content">
-    <strong class="govuk-tag govuk-phase-banner__content__tag">prototype</strong>
-    <span class="govuk-phase-banner__text">
-      This is a functional prototype of the new proposed eForm CRM7.
-    </span>
-  </p>
-</div>
