--- app/views/old_layouts/application.html.erb
@@ -1,31 +1,0 @@
-<% content_for?(:page_title) ? yield(:page_title) : fallback_title %>
-
-<% content_for(:head) do %>
-  <%= csrf_meta_tags %>
-  <%= csp_meta_tag %>
-<% end %>
-
-<% content_for(:service_name) do %>
-  <%= link_to service_name, root_path,
-              class: 'govuk-header__link govuk-header__service-name', id: 'header-service-name' %>
-<% end %>
-
-<% content_for(:header_navigation) do %>
-  <% render partial: 'layouts/header_navigation' %>
-<% end %>
-
-<% content_for(:phase_banner) do %>
-  <% render partial: 'layouts/phase_banner' %>
-<% end %>
-
-<% content_for(:content) do %>
-  <%= yield %>
-<% end %>
-
-<% content_for(:footer_links) do %>
-  <% render partial: 'layouts/footer_links' %>
-<% end %>
-
-<%# render partial: 'layouts/developer_tools' if FeatureFlags.developer_tools.enabled? %>
-
-<%= render template: 'layouts/govuk_template', layout: false %>
