--- app/views/old_layouts/mailer.text.erb
@@ -1,1 +1,0 @@
-<%= yield %>
