+++ gems/laa_multi_step_forms/app/helpers/laa_multi_step_forms/application_helper.rb
@@ -3,5 +3,23 @@ module LaaMultiStepForms
     def current_application
       raise 'implement this action, in subclasses'
     end
+
+    def title(page_title)
+      content_for(
+        :page_title, [page_title.presence, service_name, 'GOV.UK'].compact.join(' - ')
+      )
+    end
+
+    # In local/test we raise an exception, so we are aware a title has not been set
+    def fallback_title
+      exception = StandardError.new("page title missing: #{controller_name}##{action_name}")
+      raise exception if Rails.application.config.consider_all_requests_local
+
+      title ''
+    end
+
+    def service_name
+      t('service.name')
+    end
   end
 end
