+++ gems/laa_multi_step_forms/config/locales/en/helpers.yml
@@ -1,0 +1,17 @@
+---
+en:
+  helpers/dictionary:
+    YESNO: &YESNO
+      'yes': 'Yes'
+      'no': 'No'
+
+  helpers:
+    back_link: Back
+    start_again: Start again
+    submit:
+      save_and_continue: Save and continue
+      save_and_come_back_later: Save and come back later
+      save_and_submit: Save and submit application
+      find_address: Find address
+      try_again: Try again
+      apply_in_eforms: Apply in eForms
