+++ gems/laa_multi_step_forms/spec/dummy/app/controllers/application_controller.rb
@@ -1,2 +1,2 @@
-class ApplicationController < ActionController::Base
+class ApplicationController < LaaMultiStepForms::ApplicationController
 end
