+++ gems/laa_multi_step_forms/spec/dummy/app/controllers/home_controller.rb
@@ -1,0 +1,5 @@
+class HomeController < ApplicationController
+  def index
+    render text: 'Home'
+  end
+end
