+++ gems/laa_multi_step_forms/spec/dummy/config/locales/en.yml
@@ -1,33 +1,21 @@
-# Files in the config/locales directory are used for internationalization
-# and are automatically loaded by Rails. If you want to use locales other
-# than English, add the necessary files in this directory.
-#
-# To use the locales, use `I18n.t`:
-#
-#     I18n.t "hello"
-#
-# In views, this is aliased to just `t`:
-#
-#     <%= t("hello") %>
-#
-# To use a different locale, set it with `I18n.locale`:
-#
-#     I18n.locale = :es
-#
-# This would use the information in config/locales/es.yml.
-#
-# The following keys must be escaped otherwise they will not be retrieved by
-# the default I18n backend:
-#
-# true, false, on, off, yes, no
-#
-# Instead, surround them with single quotes.
-#
-# en:
-#   "true": "foo"
-#
-# To learn more, please read the Rails Internationalization guide
-# available at https://guides.rubyonrails.org/i18n.html.
 
 en:
-  hello: "Hello world"
+  service:
+    name: Claim a non-standard magistrate fee
+
+  flash:
+    success:
+      title: Success
+    notice:
+      title: Notice
+    alert:
+      title: Alert
+
+  layouts:
+    header_navigation:
+      menu_button: Menu
+      nav_aria_label: Menu
+      button_aria_label: Show or hide menu
+      your_applications: Your applications
+      sign_in: Sign in
+      sign_out: Sign out
