+++ gems/laa_multi_step_forms/spec/dummy/config/routes.rb
@@ -1,3 +1,5 @@
 Rails.application.routes.draw do
   mount LaaMultiStepForms::Engine => '/laa_multi_step_forms'
+
+  root 'home#index'
 end
